Modify i2c init script
Change command from modprobe to insert and wait for confirm remove or
delete

Dut to we went to fix i2c busses number, but we still could not fix
the i2c busses number by modprobe, so we used insmod to insert modules.
We will wait 3 seconds and record log in syslog.

Signed-off-by: Eddy Weng <eddy.weng@mic.com.tw>